+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Removed support for LXT, LXT2, VZT, and AET2 formats.
 - Removed "Signal Search Hierarchy" dialog.
 - Removed setting to disable the "Alternate Wheel Mode".
+- Removed legacy VCD loader.
 - Removed rc variables:
     - `alt_wheel_mode`
     - `atomic_vectors`